requirements.txt
boto3>=1.34.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
tabulate>=0.9.0
python-dotenv>=1.0.0
